refactor(admin): extrait AdminResponsables/AdminReferentiels via FormProvider + suspense
- FormProvider + useIndicateurFormContext (wrapper typé) pour récupérer le form
  depuis le contexte plutôt que de threader control/register.
- Sections référentiels & responsables extraites dans leurs propres composants,
  chacun avec sa query.
- Queries 'all-for-select' converties en queryOptions partagés, préchargés via
  ensureQueryData dans les loaders ($id + nouveau) et consommés en
  useSuspenseQuery (plus d'état de chargement dans le select).
